Fix contributors page rendering for Safari
I noticed that the contributors page wasn't rendering correctly in
Safari on my iPhone -- the first team member's face/name show ups, and
then after that there is only white, no other team members or footer.

I think the issue is that there are multiple, conflicting strategies for
"clipping" the profile photos into hexagons, and there's some kind of
clash there.

With this change, the page renders correctly for Safari and Chrome. On
Firefox, the page will render, but it no longer has hexagon-clipped
profile photos. According to http://bennettfeely.com/clippy/ the
clip-path property doesn't work in Firefox, although the previous
implementation did manage to get it working by embedding an svg in the
HTML and then loading that.

IMO losing the hexagon-clipping on Firefox is worth it for the page to
render in Safari, although admittedly that calculus is subjective.

(I also haven't checked to see how other browsers are affected.)