Make Heroku upgrade info match blog post
The guide on the website has a hanging "by running:" followed by nothing which is confusing. Someone already noticed and fixed this in #431 but those didn't get translated to the upgrade guide on the website. I've copied the entire question wording verbatim for consistency, but if someone thinks it'd be better not to include the note about support, question change, etc. I can update this. At the very least, the "by running" should be removed so the sentence makes sense.